Fix painting to use correct positions
I fixed(?) the viewport scaling, at the cost of the game world is now stretched to fit in the display size which is wrong.  The game should pick a point to center and then draw as much of the game around that center point as it can at undistorted scale.